refactored and fixes
