+++ src/javascript/app_2/App/Components/Elements/AccountSwitcher/index.js
@@ -1,0 +1,1 @@
+export * from './account_switcher.jsx';
