+++ src/javascript/app_2/App/Components/Layout/Header/index.js
@@ -3,4 +3,3 @@ export * from './login_button.jsx';
 export * from './menu_links.jsx';
 export * from './toggle_menu_drawer.jsx';
 export * from './toggle_notifications_drawer.jsx';
-export * from './upgrade_button.jsx';
